Removed the boolean based forking.
What
===
Removed the boolean based forking and replaced it with two separate
functions that get called for calculating the checksum modifier.

Why
===
Passing in an argument to fork requires another if decision, where-as at
that point in the flow the code already knows if it's odd or even and
it is more efficient to not work again.