Optimise loop by removing string reversal
The previous version reversed the string and then iterated through
the digits (in effect from right to left of the original string).

The reversal is slow (particularly on longer strings), this version
removes the reversal and iterates from left to right.

However, to do this correctly it must test whether the string is an
odd number of digits in length and deal with the first digit if the
string is an odd length.